requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
sqlparse==0.5.0
typing_extensions==4.11.0
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
